Make content a phony target
Running the notebooks might fail, which means `make content` sometimes
has to be run repeatedly even though src hasn't changed in the meantime
and content already exists. So it doesn't make sense to check staleness
based on the state of the src and content directories. And
src2content.sh implements its own granular staleness check of each
notebook, so it's unnecessary to boot.